Break absolutely everything.... Well, hopefully not.
This refactors every extension that I currently build under MSVC to use the new HHVM_RC* macros to register constants. This also removes any global constants that are the same as a define that's already provided, and also removes a few global constants that are never referenced from the C++ side, and registers the constant to a literal value instead.