More trailing commas in generic parameter lists
Summary: yacc grammar supports them, but there is a logic in Scanner that duplicates it, and does it wrong.

This has the same problems as https://github.com/facebook/hhvm/issues/5752, where we do generate them from hh_format, but existing versions of HHVM won't parse them. The comment there says "hh_format should remove it until 3.6 is no-longer supported", which apparently is "in March 2016." I'll look into it, but regressing hh_format feels super weird, since it doesn't have such lexer hacks and a type list looks the same to it regardless of position... we can just disable trailing commas everywhere in typehints, if that's whats necessary.

There are also more comma related things in this file (for shapes and anonymous function parameter lists) that probably don't work as well - also an issue for separate diff.

Reviewed By: jwatzman

Differential Revision: D2314177

fb-gh-sync-id: 9ca70b15465c6a6b73f7542074ff044ee9fd5c17