Revert "Revert "Fix yield tests""
Summary: This is the autopriming diff again, along with the other fixes that were reverted. The underlying issue
was that the bytecode emitted for the generator extension depended on a runtime option. For this to
work the systemlib md5 for extensions needs to be mangled. Additionally HHBBC needs to understand
the runtime options used to mangle these md5s so that it loads the appropriate units.

This reverts commit a28128b36e5c4098b36248fac754fbaf574f4705.

Reviewed By: jano

Differential Revision: D2569060

fb-gh-sync-id: 80b50ebc49db412d588052b78a0adcdc7e5fc1b9