Make hh_client restart more reliable / helpful
Summary: I imagine it's really annoying to be stuck with a broken hh_server (and we've
been breaking it a lot lately), being told to restart it, then having the
suggested command fail with no explanation. Improvements:

* Wait a little longer after sending the kill signal before concluding that
  server won't die
* Suggest pkill when that happens

Incidentally, I believe that until recently we had a bug with restart:
hh_server would write to the PidLog before it tried to acquire the file lock,
so calling `hh_client start` when a server was already running would cause the
PidLog to be overwritten and subsequent kills to fail. However, given that the
lock has been moved to an earlier point in startup (see D2543288), this may no
longer be an issue.

TODO: Log to Scuba when a restart fails

Reviewed By: dabek

Differential Revision: D2569862

fb-gh-sync-id: f119dd7029caf1ba4ce6c5d6079a44a5f409d427