A couple of misc. fixes for MSVC
The first is due to a compiler bug causing it to attempt to construct `ALocBits{meta->conflicts}` as a `const` value. Due to the fix for this likely being a breaking change (in MSVC compiler itself), the fix for that won't be in Update 1, so just remove the `const` for now.
The second is due to a random parser issue, and the original also didn't comply with the normal code style standards of HHVM to begin with, so just put curly brackets on the loop and the problem is solved.