Rename use_watchman to use_watchman_2
Summary: I'd like to enable watchman again next week, but only for the new Hack release.
Setting `use_watchman = true` would enable it for old releases as well, which
is not why I want.

This solution is clowny as hell, and IMO the proper fix is to use semver. It's
a work in progress here: {D2576911}, but it doesn't build with Buck yet.

Reviewed By: dabek

Differential Revision: D2576951

fb-gh-sync-id: 6624adf25196aab1d8a0fb9dd58314110af00763