Revert: [hhvm][oss] Support using fb-mysql instead of webscalesqlclient
Summary: We need to revert this. As it is, people building HHVM will always see this:

```
modified:   third-party (modified content)
```

and the reason for that is that we need to add the following
files to the `mysql-5.6` `.gitignore` file:

```
Docs/INFO_SRC
libmysql/merge_archives_webscalesqlclient.cmake
libmysql/webscalesqlclient_depends.c
```

drilibo change apparently causes these files to be added during
build-time.

So instead of committing this and having that problem and having
to re-commit again, let's revert this. Change the `.gitignore` in
`mysql-5.6`, update the `webscalesqlclint` SHA in `hhvm-third-party`
and then the `hhvm-third-party` SHA again like we did with this diff.

Reviewed By: chipturner

Differential Revision: D2578951

fb-gh-sync-id: 8b9564aed5ccd3a35d5082571abedb5a1c6d9377