Fix the fallback in fls64
`folly::findLastSet` returns a base-1 bit index, we need a base-0 index. `folly::findFirstSet` already returns a base-0 index, so there is no problem there. This was actually causing HHVM to never finish saving the registers in the vasm unique stubs, and causing it to attempt to append an infinite number of pop instructions.