Eliminate most uses of runtime/base/zend-php-config.h
`zend-php-config.h` doesn't belong in the core runtime, as it's not even remotely portable to other platforms, nor are most of the defines in it even correct.

It was only used for 3 things, logical_filters in ext_filter, it was included in ext_sockets, but not actually used, and ext_gd uses it extensively. This only removes the ext_filter and ext_sockets use of it. ext_gd's dependence will be dealt with in a separate PR.