Optimize LdMIStateAddr
Summary: vasm-copy is hardcoded only to handle rvmfp().  Change it such that it also
handles rvmtl(), then we can fold copies from the lea we emit for LdMIStateAddr
as well.  This needed a another change to vasm-simplify to reduce copyargs to a
series of copys (when we know the sources never appear as dests), because
vasm-copy ignores copyargs and only works on copy.  We end up generating
copyargs quite frequently for rvmtl() because minstrs usually end up running
minstr helpers.

Next, disable GVN for LdMIStateAddr.  This makes us emit way more
LdMIStateAddrs, but the new folding in vasm-copy causes us to emit no codegen
for them.

Finally, change emitRatchetRefs()/ratchetRefs() such that they emit a
LdMIStateAddr for their phijmp.  This way we can DCE away the lea from the
LdMIStateAddr that is created at the start of our ratchet block.

Reviewed By: swtaarrs

Differential Revision: D2569695

fb-gh-sync-id: 8c5b0f2ce663e5635b39b3a69def606177f11bae