Fix crash when compiling Drupal 8 repo auth (and prob others)
Summary: `std::sort` requires its comparator to be a strict weak
ordering, otherwise undefined behavior, such as an out-of-bounds read,
occurs. This comparator was subtly not: the `a.compare(b)` function
returned `-1` if `a < b`, `+1` if `a > b`, and `0` if `a == b` -- but
`-1` and `+1` both coerce to boolean `true`, and the comparator is
supposed to return true iff `a < b`. This very occasionally lead to a
crash, due to an out-of-bounds read off one end of the vector, easily
visible with ASAN. Actually triggering it is somewhat finicky; the same
code always triggered it, but I haven't found a small example; the
`phpunit` subdirectory of Drupal 8 is the best we found.

Fixes https://github.com/facebook/hhvm/issues/6296

Reviewed By: paulbiss

Differential Revision: D2587046

fb-gh-sync-id: 9cc695b899689ec2bfa9e2a0d5bdca6844785ba1