Implement `getReturn` on Generators
Summary: This diff implements the `getReturn` API as described in
`https://wiki.php.net/rfc/generator-return-expressions`. It adds a new method
to the Generator object called `getReturn` which returns the generators return
value. For unfinished generators (or generators that have thrown an exception)
calling this method will throw an exception. For generators that don't return a
value, calling `getReturn` will return `null`.

Implementation wise, we store the return value in the same cell as the yield
value. This leads to some odd behavior with methods like `current()`, so we
need to branch based off the generators current state in the implementation of
that method.

Fixes: https://github.com/facebook/hhvm/issues/6057

Reviewed By: paulbiss

Differential Revision: D2522421

fb-gh-sync-id: 2e32b8243fe9abfd98c1a7d61ee9b9bfbaf6f7f9