Fix an unwinder bug
Summary: If exception_handler catches a php exception, and the catch variable
is the last reference to the exception, it will be destroyed at the
end of the catch block, possibly calling __destruct methods. But
unwindPhp leaves vmpc()==nullptr and vmfp()==nullptr, so we're not in
a valid state for re-entry. So we need to fix that before the end of
the catch block; but in addition, the callers use vmpc() == nullptr as
a stop condition, so we need to reset vmpc() before returning.

Reviewed By: mxw

Differential Revision: D2594143

fb-gh-sync-id: a29c085f6856a236b69abc5f63d06cb09ae0c3ad