Pass base directly to {Empty,Isset,Set}Prop when we know it's an Obj
Summary: This only affects new-minstrs, old-minstrs already do the right thing.

We were always passing the pointer to the base for {Empty,Isset,Set}Prop, but
there are cases where we have the available value for the base and we know it's
an object, so we should pass that instead.

Reviewed By: swtaarrs

Differential Revision: D2587285

fb-gh-sync-id: 4fae3ef00809ef5bbd879c3a59f6f2c3bcbe95ab