Add ElemArrayU IR opcode, tighten return type of ElemArrayD
Summary: ElemArrayU implements ElemU when we know the base is an array or a boxed array.
This gives us the benefit of going straight to the ElemUArray helper instead of
going through ElemU.  I also changed ElemUArray not to take &tvRef as an
argument, so ElemArrayU also lets us skip another use of MIState.

As I was adding memory-effects for ElemArrayU, I noticed that ElemArrayD's
return type was PtrToMembGen, but that's too broad.  ElemArrayD will always
return the address of an array element.

Also add a main() function for two minstrs tests in test/quick.  They were
running completely in pseudo-mains, but we generally just punt back to the
interpreter on minstrs in pseudo-mains so we weren't actually testing the JIT.

Reviewed By: swtaarrs

Differential Revision: D2589439

fb-gh-sync-id: c66b1147833015e1abf619b0a3ff964f9f8c06c9