Delete HHIR relaxGuards pass
Summary: This pass has been replaced by region-level guard relaxation and it's
currently off. In theory, the fact that HHIR guards don't relax
anymore should also allow us to get rid of the checks for
typeMightRelax in IRBuilder and simplify.cpp, but my initial attempt
to do so along with this diff caused a performance regression.  So I'm
keeping typeMightRelax and its uses for now and I'll follow up on it
later.

Reviewed By: alexmalyshev

Differential Revision: D2600733

fb-gh-sync-id: 8071abaaeb7f2b9eb462ed3d9897dbb871795eec