Remove CodeModel
Summary: CodeModel is dead code.
This is largely a revert of D1027004 plus or minus for
the little changes/additions/deletions since then.

Of note is the change to the PHINQ executeQuery() call wherein a serialized CodeModel of the query's AST was previously passed as the first argument to the function.  Since CodeModel is now gone, I've modified that to always pass NULL.

Obviously this will break any code relying on having a valid code model there, but as there's no current evidence that CodeModel is in use (nor PHINQ), it's probably a safe thing to do pending removal of PHINQ as well.

Reviewed By: JoelMarcey

Differential Revision: D2598919

fb-gh-sync-id: eea272cc8b7c674ba0866b09194537136a12465d