Remove PHINQ
Summary: As with CodeModel, this represents a dead feature with no users,
few people who understand it, and is a source of bit rot.

This is largely a revert of D1088496.

Of most notable difference is that async short lambdas were added
after this was put in place, and I had to reorder the
`lambda_expression` parser rule to give precedence to async lambdas.
Otherwise, the labmda would resolve first, then the T_ASYNC would
be orphaned and regard the lambda as unexpected.
I plan to refactor this rule in a later diff to both avoid that
conflict and reduce copypasta in the parser itself.

Reviewed By: JoelMarcey

Differential Revision: D2599513

fb-gh-sync-id: 6b3a083f4468df2673eccb32523410b92007eb2f