Repurpose Vxls for the register allocation phase only
Summary: Stop using the Vxls object for all of XLS, and instead use it only for
the register allocation phase---the only phase which actually needs a
large number of working data structures.

Also some very minor refactoring and a bunch of documentation and inline
comments that helped me figure out what was going on in-depth.

Reviewed By: edwinsmith

Differential Revision: D2589769

fb-gh-sync-id: f6cc8afaefc03c9955e0cc5043745a9c1d8cef1b