More whack-a-mole with incorrect line numbers in positions
Summary: Where running fixme.py script I've found one error that it was not fixing, and the reason was incorrect line numbers. This is another case of Lexer_hack.back not supporting backtracking through tokens that contain newlines.

The truly right fix for that would be to replace OCamls built in Lexing functions and types with our own wrapper that would track more information about previous lexeme and allow to implement Lexing.yyback correctly supporting backtracking across the line breaks. Doesn't seem to be worth it now -  for now I just plugged the holes we know of, as we did before.

Reviewed By: int3

Differential Revision: D2612095

fb-gh-sync-id: 0a110646ca929226fa41064ebe9787d5f9859acf