Use LowPtrs for some common TCAs
Summary: This saves some amount of memory in production. I added template
arguments to AtomicLowPtr for specifying the memory ordering, since
SrcRec::m_topTranslation needed stronger guarantees than what AtomicLowPtr was
providing.

We could unconditionally use 4-byte offsets for the code addresses in SrcRec,
but that would be more complex and we don't really need the memory in
non-lowptr builds. It would also require additional instructions to load the
value, and I know we've been sensitive to how we get m_topTranslation in the
past. We can still explore doing that in the future.

Reviewed By: mxw

Differential Revision: D2611752

fb-gh-sync-id: edf2934ca45b343e12ff770d57c9d279f12e7a36