Remove the global LibIConv requirement
Summary: LibIConv is currently required globally, but only `ext_iconv` and `ext_gd` use it.

This was originally opened as #5723 (D42495), but old age caught up with it, and it ended up being closed before it was merged.

Due to other changes, the only pieces of this PR that were left being needed are the actual removal of the global check, and, apparently, also the removal from the list of libs that are implemented globally in the extension config mechanism.
Closes https://github.com/facebook/hhvm/pull/6415

Reviewed By: sgolemon, JoelMarcey

Differential Revision: D2601445

Pulled By: sgolemon

fb-gh-sync-id: 7e5feac644b06e197f4d511062c06a108b268cef