A couple of misc. fixes for MSVC
Summary: The first is due to a compiler bug causing it to attempt to construct `ALocBits{meta->conflicts}` as a `const` value. Due to the fix for this likely being a breaking change (in MSVC compiler itself), the fix for that won't be in Update 1, so just remove the `const` for now.
The second is due to a random parser issue, and, as far as I know, the original also didn't comply with the normal code style standards of HHVM to begin with, so just put curly brackets on the loop and the problem is solved.
Closes https://github.com/facebook/hhvm/pull/6419

Reviewed By: edwinsmith

Differential Revision: D2601502

Pulled By: sgolemon

fb-gh-sync-id: 6867f05d1dffe15d72b55c34fcb3a18c5e06c42f