Fix the fallback in fls64
Summary: `folly::findLastSet` returns a base-1 bit index, we need a base-0 index. `folly::findFirstSet` already returns a base-0 index, so there is no problem there. This was actually causing HHVM to never finish saving the registers in the vasm unique stubs, and causing it to attempt to append an infinite number of pop instructions.
Closes https://github.com/facebook/hhvm/pull/6422

Reviewed By: sgolemon, JoelMarcey

Differential Revision: D2601526

Pulled By: sgolemon

fb-gh-sync-id: 0cac1f7467b14d11a6ad5fa2f098f0ed30942922