Document the Map hhi for use in our documentation.
Summary: Collections are in process of being converted to HNI, but in the meantime we need
documentation for Collections for the new docs. So we document the HHI instead.

This fills in the documentation gaps for Map. We can then transfer this documentation
over to the HNI when that lands.

Notice too how I add guide tags to some of the docblocks. These will be links to
appropriate guides where appropriate.

If we like this approach, I will do the rest of the collections in this form.

Reviewed By: fredemmott

Differential Revision: D2612775

fb-gh-sync-id: 5651617e24803b8f10238a3d5cc49ba0b9fb561c