Document the ImmXXXX hhi for use in our documentation.
Summary: Collections are in process of being converted to HNI, but in the meantime we need
documentation for Collections for the new docs. So we document the HHI instead.

This fills in the documentation gaps for `ImmVector`, `ImmMap`, `ImmSet`.
We can then transfer this documentation over to the HNI when that lands.

Notice too how I add `guide` tags to some of the docblocks. These will be links to
appropriate guides where appropriate.

Reviewed By: fredemmott

Differential Revision: D2614770

fb-gh-sync-id: 3f1c5ccb8f5dc041dc4f53f921a219528dd4137a