Document the Pair hhi for use in our documentation.
Summary: Collections are in process of being converted to HNI, but in the meantime we need
documentation for Collections for the new docs. So we document the HHI instead.

This fills in the documentation gaps for `Pair`. We can then transfer this documentation
over to the HNI when that lands.

Notice too how I add `guide` tags to some of the docblocks. These will be links to
appropriate guides where appropriate.

Notice too that I added `internal` to the constructor ... this shouldn't be exposed in
the docs.

Reviewed By: fredemmott

Differential Revision: D2617682

fb-gh-sync-id: dab1820894cea4fc6a1d7f0d40f913d635e98332