Don't emit FPushFuncD for "Class::Method" strings
Summary: FPushFunc support was added in D2514895.  There's no support for FPushFuncD
handling these strings in the interpreter or the jit.  As such,
test/slow/function/variable-call-method.php fails when run under repo mode.  I
managed to fix it in the interpreter, but I got worried about jit-perf so I
dropped the diff and did this instead.  This shouldn't affect perf as nobody
uses this in www.

Fix the bytecode docs along the way.

Reviewed By: swtaarrs

Differential Revision: D2620006

fb-gh-sync-id: 19bb5855d2de6b2108a837eaf9bdc6d054b9e07c