Document the Set hhi for use in our documentation.
Summary: Collections are in process of being converted to HNI, but in the meantime we need
documentation for Collections for the new docs. So we document the HHI instead.

This fills in the documentation gaps for Set. We can then transfer this documentation
over to the HNI when that lands.

Notice too how I add guide tags to some of the docblocks. These will be links to
appropriate guides where appropriate.

Reviewed By: fredemmott

Differential Revision: D2613982

fb-gh-sync-id: 8776dc3f53b8e1e80b3fc1bcd9d8c7c1296db92e