Complain loudly if Hack is not able to honor the --ai flag
Summary: When there are any errors in the code being analyzed by Hack, it does not invoke the AI code. This can be disconcerting when running in server mode since the failure is silent and only becomes apparent when hh_client commands do not work as expected.

We now complain loudly when --ai is specified but the AI code is not invoked.

Reviewed By: int3

Differential Revision: D2619950

fb-gh-sync-id: f1f6062c688ed60e933159ce5bf133573b044e91