Fix assert_fail in MixedArray::CreateVarForUncountedArray()
Summary: MixedArray::CreateVarForUncountedArray() delegates to
MixedArray::MakeUncountedPacked() when its input is a PackedArray.
MakeUncountedPacked() asserts that the PackedArray is not empty, as
CreateVarForUncountedArray() is expected to handle empty arrays.  That check was
wrong, it was doing a pointer comparison with the static empty array, but we
were passing in an empty packed array.

Reviewed By: markw65

Differential Revision: D2621483

fb-gh-sync-id: e95691b870611ea52d0657618a76e488eecee478