Fix Mac hotprofiler build
Summary: These constants were previously IDL defined as hardcoded
numbers, but are now properly defined with the system values. This is
problematic on OS X where three of them don't exist, so `ifdef` around
it. (`CLOCK_REALTIME` does exist on OS X, but I define it conditionally
anyways for consistency.)

Reviewed By: paulbiss

Differential Revision: D2620747

fb-gh-sync-id: 418ca481a9e12c6cdd58548d6bc92b617c572afd