Don't emit struct arrays for collection initializers
Summary: Right now, HashCollection holds onto a MixedArray and it assumes the array it is
constructed with is a MixedArray.  This crashes when we pass it a StructArray.
We could make the HashCollection ctor escalate the StructArray to a MixedArray,
but that's just wasted work at runtime.  Make the frontend emit a MixedArray
always for arrays being used for map/set initializers.

Change the array emission logic to emit NewStructArray regardless of the value
of RuntimeOption::DisableStructArray.  The VM already knows to check the option
and will create a mixed array when struct arrays are disabled.

Reviewed By: swtaarrs

Differential Revision: D2616409

fb-gh-sync-id: 1674b06fcfb1434089fe745e2b629e763b2cc331