Perform dirname at compile time when possible.
Summary: If the argument to dirname() is a string literal (or equivalent), we can call dirname() at compile time and just return the precomputed value at runtime. This is already done for ini_get().

This makes
 dirname(__FILE__) . "/some_relative_file"
as efficient as using a hardcoded absolute path.

I don't see a measurable performance improvement with WordPress, but I don't see a downside to this optimization.
Closes https://github.com/facebook/hhvm/pull/6456

Reviewed By: sgolemon, JoelMarcey

Differential Revision: D2605980

Pulled By: sgolemon

fb-gh-sync-id: f770327f036c4588cfb7957652c11750dc42f369