Update for VS 2015 Update 1 RC
This removes all of the constexpr offsetof changes, the constexpr bitfield constructor changes, and also switches the array member size workarounds to a form that is compliant with the C++ spec, as the current form isn't even remotely close to being compliant.
This also switches the auto templated overload version of registerNativeDataInfo to be the only version, as it is far more readable. This also updates it to reflect that a check for `scan` was also added.
The rest of the auto templated overload workarounds have been left, as has the `std::chrono::duration<double>` add workaround, because the template overload issues are cleaner without the workarounds, and the double workaround is lossy.