Rejigger jit::Ptr types
Summary: The PtrTo* bits in m_bits are gone, and the entire pointer component
of a type lives in m_ptrKind, which is a proper bitmask. Ptr has a real Top and
Bottom now instead of Unk, which was mostly Top but also weird. This frees up a
number of bits and has some consequences on the types we can and can't
represent.

Because the pointer component of a type applies equally to all bits in Gen, we
can no longer represent types like `Str|PtrToInt`. `TStr | TPtrToInt` will now
return `Int|Str|PtrToInt|PtrToStr`. I'm not aware of any situations where we
actually use a type like that so I don't expect it to be a problem. There
wouldn't even be any way to distinguish between the different components at
runtime.

In exchange, we gain much more precision for Ptr kinds. We can now represent
things like `PtrTo{Prop|Elem}Gen`, or `PtrTo{Frame|Stk}Gen`. This should help
with the dest types of some phis, which may help memory-effects.

The extra space in m_bits means there's now room for UncountedStr and
UncountedArr. This will let us store KindOfStaticString again in the jit, which
I'll try in a separate diff to isolate any performance impact.

Finally, I renamed Ptr::Arr to Ptr::Elem because nobody has time for crap like
PtrToArrArr.

Reviewed By: mxw

Differential Revision: D2629380

fb-gh-sync-id: a1595e806d711132f24b0cdd608d51a70050b724