Fix collection_ini_options test on OS X
Summary: When you append a symlink to an INI option, we expand so that both the symlink and its target get appended. This test appended `/tmp`, which is typically a symlink on OS X but not on Linux, leading to inconsistent output and a failed test. Let's just use some random directory that won't exist anywhere.

Reviewed By: JoelMarcey

Differential Revision: D2639051

fb-gh-sync-id: 7999f7d69c57552d22a5900bc4d1821e8d7d3269