Fix canonicalization of paths that start with `.`
Summary: The path canonicalization function was trying to drop the
superfluous `.` in `foo/./bar`, but dropped a leading `.` too, which is
wrong. This was previously largely covered up by `TranslatePath`
treating an empty string as CWD, but that was recently fixed in
D2437818.

One place in RID also has some code deleted in this change -- due to the
bad canonicalization, it would never have triggered, but now it does...
and crashes since `g_context` hasn't been initalized yet.

Fixes https://github.com/facebook/hhvm/issues/6402

Reviewed By: FBNeal

Differential Revision: D2620909

fb-gh-sync-id: 41b9847cd103f145e22406d83b60f10cfe11c0f8