More efficient vasm for ConvStrToBool
Summary: Linker script tells us StringData::toBoolean() is hot, which makes me think the
calls are coming from ConvStrToBool.  We can implement that without C++ helpers.

Strings in PHP are silly and convert to false when they are equal to "" or "0".
The case for "0" is expensive and should be unlikely, so stuff that in our cold
section.

Reviewed By: binliu19

Differential Revision: D2638139

fb-gh-sync-id: edc2c84fe300ae309878ea529578fe0800f6fb77