Add is_subclass_of() case to irgen-builtin
Summary: It's almost the same as is_a(), except it returns false when the LHS and the RHS
have the same class.

Added a new IR opcode EqCls to compare two arbitrary TCls values for equality.
We had ClsNeq but that assumed one of its inputs was a constant value.  I just
replaced it with EqCls, as we can still emit good code for two runtime class
values.

Reviewed By: paulbiss

Differential Revision: D2634831

fb-gh-sync-id: c009ca2d811909de6f46f9c83ca25b15e7775cc9