Add new MethodExists IR opcode
Summary: As the name says it implements method_exists(), but it takes a TCls and a TStr
instead of a TObj|TStr and a TStr.  Right now we only lower a call of
method_exists($a,$b) to MethodExists when we know that $a is an object and $b is
a string.  Getting the $a <: TStr case was more effort than this is worth.

Reviewed By: swtaarrs

Differential Revision: D2627980

fb-gh-sync-id: 681b6298e43e8cdf836d0e4ffeede92b7ace08c5