Add a profile-count field for IR Blocks, and use it for code layout
Summary: This diff adds a `profCount' field to IR blocks, which is filled in
during IR generation.  For the outermost region being translated,
these counters are the profile counters of their corresponding region
blocks.  For inlined regions, the profCounts are scaled based on the
ratio of the calling block's profile counter over the callee's entry
counter.

This diff also uses this new profCount field in the Vasm code layout
pass, instead of the profile-translation counters.

Reviewed By: mxw

Differential Revision: D2635200

fb-gh-sync-id: ca6c6a0bf7c8bc100839a57f67aa1005fddb31c1