Group use declarations
Summary: Not super pretty, since the `function` or `const` modifier can
appear inside or outside, and tracking that in our AST is a PITA.

We differ in one place: we admit `use Blah\{\Foo}`, whereas PHP7 rejects
it (the second `\` is invalid). Not super hard to fix, but it continues
a combinatorial increase in our parser rules, so I didn't bother.

Fixes https://github.com/facebook/hhvm/issues/6056

Reviewed By: sgolemon

Differential Revision: D2561610

fb-gh-sync-id: 33ec874b469a52747e1a2a68ca721e1c63fbfefd