Convert ext_gettext to HNI
Summary: This converts the very simple gettext extension from `ext_zend_compat` to use HNI instead.

I originally did the port because of me failing to realize where libintl was being searched for, but this is a simple enough extension that I didn't see any reason not to submit a PR to switch it over.
Closes https://github.com/facebook/hhvm/pull/6279

Reviewed By: sgolemon, JoelMarcey

Differential Revision: D2464494

Pulled By: sgolemon

fb-gh-sync-id: bea80071918ed6ea1da5e119ee90a192a2f6a24f