get (line, col) pair from Pos
Summary: We have a bunch of access functions in `Pos`, but none of them
actually grab the line and column for both the start and end
of the token range.

Since the 'get the lexer token' functions are deprecated, add
something that returns two (line, col, line, col) without leaking
implementation details.

Reviewed By: int3

Differential Revision: D2646421

fb-gh-sync-id: 3233de86a9106aa4b35c2490a2b171d8249ada57