Convert closure to HNI
Summary: Standard conversion for the most part.
I did make Closure somewhat special by putting its HNI definition
at the top of the main systemlib rather than std's extension-slib.
This ensures that it is the very first thing loaded and hoisted
into the runtime, and therefore closures are now usable in all
systemlibs.
This also reduces the number of tiny "special case" exceptions we
need throughout the runtime to accomodate for closure being one
of the last things to bind (via IDL).
Instead, we get one contained exception in that the binding for
closure happens in moduleLoad() rather than moduleInit() since
the main systemlib is loaded between these two callbacks.

Reviewed By: JoelMarcey

Differential Revision: D2649934

fb-gh-sync-id: 9fc93212ebf7817e78a4b930c4412ae2bd8a5828