Document MCRouter::result() versus MCRouter::unserialize() and fix mcr_void
Summary: It came up in review of D2646622 that the data dance going on in
result/unserialize is unintuitive.  Add some comments.

I also noticed that mcr_void<op> had mc_op_version hard coded
when it should be for any generic op which doesn't take any
arguments.  This works fine atm as only version() is called
this way, but it should be future proofed for other methods.

Reviewed By: JoelMarcey

Differential Revision: D2650092

fb-gh-sync-id: 79338f9c4e5ff75f563c88ffda151cd637bbc76b