Improve profile accuracy of DecRefs release-path placement for Ret{C,V}
Summary: The profiling for whether the DecRefs take the release path or not
were keyed by the BC profTransID, BC offset, and type.  That's
generally enough, except for Ret{C,V}, which may decref many locals in
line.  This diff extends that key with an optional 'local id' to
disambiguate things and make the profile more accurate.

Reviewed By: paulbiss

Differential Revision: D2638370

fb-gh-sync-id: c350073aeea36cd1ec17a07aa72a72b777599317