Fix `??` operator on falsey values
Summary: `$a ?? $b` should return `$b` only if `$a` is unset or null; we were also doing it on any falsey value. Also had to fix a bug in the JIT where checking if uninit was of type null got misoptimized in the simplifier.

Fixes https://github.com/facebook/hhvm/issues/6545

Reviewed By: ricklavoie

Differential Revision: D2653756

fb-gh-sync-id: 67c0b63d2a0bb1d26395181f213ac4f617838297