Tighten up vasm unit printing
Summary: Move IRInstructions from their old position to a column on the
right-hand side, don't emit an empty line when the origin changes, and only
emit the origin instruction when it actually changes, not at the beginning of
every block. This makes the output much more compact and IMO it's much easier
to follow the flow of the vasm this way.

Reviewed By: mxw

Differential Revision: D2654095

fb-gh-sync-id: a6bfbbeb1dde9ec93956b8f36ddf8e031d08f998