Allocate RDS bits in smaller chunks
Summary: We used to allocate at least 512 bytes, ending on a 512-byte
boundary. In practice this meant we usually allocated ~950 bytes at a time for
bits. This used to make sense when we did more bit checks to see which Units
were loaded, and we might check many of them in a row, but now it just serves
to space out the rest of the data in RDS. It would be especially bad if we ever
allocated a bit that was associated with another piece of RDS data, since
they'd almost certainly not end up on the same cache line.

Reviewed By: markw65

Differential Revision: D2661504

fb-gh-sync-id: aee9e7068ca0d932c0cae6aa91240eb91904adeb