Enable PGO inlining when tracelet inliner returns a bad region
Summary: When the tracelet selector was returning a bad, non-empty region, the
JIT wouldn't try to use the PGO selector as intended.  This diff fixes
this issue by moving the call to shouldInline from
getInlinableCalleeRegion into selectCalleeRegion.

Unfortunately, doing the fix above resulted a in 1-2% performance
regression. To avoid that regression, this diff also introduces a new
maximum cost for regions generated by the PGO inliner.  Having this
separate maximum cost depending on how the region was formed is kind
of weird, but it's the only was I've found to make the first change
not be a regression so far.  This is still better than what we had
before though, so I think we should go with this for now.

Reviewed By: paulbiss

Differential Revision: D2660969

fb-gh-sync-id: a932e5ec30012c68382152219b2f78729dd2141a