Clean up InstanceDtor flags
Summary: ObjectData::InstanceDtor isn't really a flag, its a mask that
can be used to check for both kinds of native objects. Remove
the enum, add an accessor for that specific query. This makes
it plain that WaitHandle is getting both flags set without
changing any functionality, and also frees getAttributes() to
be deleted.

Reviewed By: jano

Differential Revision: D2666128

fb-gh-sync-id: 1e96557991808bb3049468af5b6d550e849e2b81