Sink IncRefs past some instructions to refine their operands' types
Summary: The pattern of loading a value from memory, increfing it generically,
and later checking its type is kind of common.  For example, an
Optimize translations is often formed by a Profile translation that
ends by pushing a value of unknown type on the VM stack, followed by
another Profile translation that guards on that type.  This diff adds
a simple pass to sink IncRefs past Check*s instructions and other
instructions that commonly occur along with them.  This can provide
additional type information that enables generating more specialized
code for the IncRef or even eliminating it completely.

Reviewed By: markw65

Differential Revision: D2648943

fb-gh-sync-id: 797dad94c4ae58d055070986771dbd91963668e5