Fixes timezone off by 8 leap minutes.
Running under Ubuntu Wily (AArch64) returns a time and timezone
offset by 8 minutes for the -5000000000 (1811) test. This is
subject to date in /usr/share/zoneinfo/US/Pacific. America/
Los_Angeles leads to a timezone of -0752, but setting the TZ to
directly to PST5PDT does not off set the timezone by the leap
minutes and returns a proper -0800.

Fixes test/slow/ext_datetime/date.php