Shrink static string table keys
Summary: Keys in the table are really LowPtr<StringData>, so we can use
a smaller key type. Since the value is an rds::Link (32-bit),
the overall table can be half as big.

Use a thread-local StringSlice to handle lookup strings. We use
a magic key that tells the eq() and hash() functions how to find
the actual string to use.

Reviewed By: markw65

Differential Revision: D2649157

fb-gh-sync-id: a02e4dfcad29d358f27b1d30c83f2f84a03d8714