Parallelize trie construction in coverage metric
Summary: This is not necessary when the only aggregation per node is adding two numbers, but with more reasons-related stuff I want to add there it will be helpful.

NOTE: in the testing I noticed that coverage metric requires an argument, but passing anything but .hhconfig root just results in a crash and "internal error"... it also doesn't attempt to do anything smart wit this path and always ends up typechecking (with hooks) the entire codebase. int3, what was the intention here?

Reviewed By: int3

Differential Revision: D2671136

fb-gh-sync-id: e1c84a4fd61a03e613aad3e4e34b4b2159c63511