Add freelists to rds allocation paths
Summary: This allows us to use the holes left by large alignments to fulfill
smaller allocation requests that come later. The current scheme is extremely
simple, but manages to keep the wasted space in normal and persistent RDS down
under 10 bytes each (it was around 9000 and 11000 before, respectively).

The maps of vectors do take up some memory but it's not memory that we have to
zero at the start of every request, and it should be far less than the 9kB *
numThreads that we save in the normal RDS.

Reviewed By: mxw

Differential Revision: D2673844

fb-gh-sync-id: 4379d83a019f3ff153e80060dbdacf7392bd274a