Optimize APCObject for persistent Classes
Summary: When the Object's Class is persistent, we can just store an array of APCHandle*
corresponding to the declared properties. If the Object has dynamic properties,
its dynProp array is stored at the end of the APCHandle* array. And in certain
cases, we can avoid the overhead of default-initializing declared properties,
constructing them in-place instead. This makes for smaller APCObjects and much
faster conversions to ObjectData.

Reviewed By: markw65

Differential Revision: D2679034

fb-gh-sync-id: 25a33ae3db829f4fda5036359df689e56e52f691