Port ext_fileinfo to MSVC
Summary: This pokes the condition around a couple of includes and defines, adds a set of `#undef`'s to config.h, and deals with the fact that `time_t` is a 64-bit value and `ts.tv_sec` is only a `long` value.

This extension actually requires quite a bit more poking of the includes, which will come as part of #6158 (D45663), due to their dependence on the upstream folly change.
Closes https://github.com/facebook/hhvm/pull/6242

Reviewed By: sgolemon, JoelMarcey

Differential Revision: D2445112

Pulled By: JoelMarcey

fb-gh-sync-id: 9332a667da4f91dad598bd9a3dcb4e9eca2efe87