Eliminate the cs_GMP_FUNC_NAME_* globals
Summary: They are pointlessly long, and, although they may be there in an attempt to prevent typos, even include a typo, as `cs_GMP_FUNC_NAME_GMP_AND` is set to `"gmp_add"`.
This just does a regex replace of `cs_GMP_FUNC_NAME_([A-Z_]+)` with `"\L\1"`. (For those who don't know, `\L` converts to lowercase in, at the very least, the regex engine used by Notepad++)

This also inlines a couple of the errors that are used in only one place, and are shorter than the name of the global.
This also removes `cs_GMP_FAILED_TO_ALTER_BIT`, which was never referenced anywhere.
Closes https://github.com/facebook/hhvm/pull/6586

Reviewed By: sgolemon, JoelMarcey

Differential Revision: D2695888

Pulled By: JoelMarcey

fb-gh-sync-id: 27aef55bf51b8243635122c0a5ff6d47174da5ec