fb_intercept segfaults when globally deregistering intercepts
Summary: The issue can reproduced with the provided test. The issue is
interleaved calls to the global deregistration fb_intercept call in both
A and B's destructor in the test. This results in the runtime trying to
destroy the same element twice, while its already in the middle of
destroying it first.

The correct approach in PHP is for both A and B to only remove the calls that
they themselves registered for interception rather than remove all of
them.

We can solve this problem in the runtime by making sure we remove the
element first before we call any destructors.

Thanks to markw65 and keyurdg for helping to diagnose.
Closes https://github.com/facebook/hhvm/pull/6515

Reviewed By: JoelMarcey, sgolemon

Differential Revision: D2621794

Pulled By: JoelMarcey

fb-gh-sync-id: 72e72a20aff2ba159fff2ebca69e4a79496d236d