Instantiation should not force localization
Summary: Localization assumes that the decl phase is done and that all classes have been
declared. However, instantiation runs during the decl phase, breaking the
assumption.

I think we used to do this localization in order to check that type constraints
were obeyed, but this is now covered in Typing_phase.localize and
NastCheck.check_happly.

Reviewed By: dabek

Differential Revision: D2701300

fb-gh-sync-id: d3bb038325d23e1974d1304cf1bb4c97262157c4