Implement PHP 7 scalar types
Summary: This is a very large diff. The most important changes are in parser.cpp, tv-helpers.cpp, and type-constraint.cpp. PHP 7 adds scalar types which can be used in two different modes, weak and strict. Strict mode is nearly identical to Hack, the sole difference is the widening conversion int -> float is permitted. Weak types are mostly analogous to the way parameter coercion currently works in PHP 5 and HHVM. The critical changes to weak types include: raising a notice for string -> number conversion with extra characters after the number, banning null conversion to number, string, and bool, and narrowing conversion from double to integer (safe conversion continue to be allowed).

In PHP 7 the caller specifies the mode of parameter type checking, and the callee decides the mode of return type checking (which was also added in PHP 7). To support this I added a bit to the ActRec flags encoding the type-checking mode. These new rules also apply to C++ builtins in PHP 7, necessitating changes to the various tvCoerceParamTo*OrThrow helpers. For PHP functions in an attempt to minimize the perf impact I added the parameter conversion to the failure path (TypeConstraint::verifyFail), where we also do Collection -> Array conversion.

In Hack mode caller preference is ignored and parameters are always strictly checked, this felt in keeping with the Hack philosophy, but if it becomes a problem it would be easy to change. For now most of these changes are gated on the new PHP7_ScalarTypes runtime option. In addition in Hack files and EnableHipHopSyntax mode we don't allow widening int -> float conversions as Hack bans this.

Finally PHP 7 is much stricter than Hack about reserving the primitive type names. When running with ScalarTypes enabled HHVM now bans primitive type names in most contexts where they had previously been allowable (in some cases even in HH mode).

For more details on the new type checking modes see https://wiki.php.net/rfc/scalar_type_hints_v5

Reviewed By: ottoni

Differential Revision: D2481743

fb-gh-sync-id: d030cdbb0db9fa6c8e4527a1ba9b2cfa0f7d34fc