Port PHP7 CSPRNG functions random_int() and random_bytes()
Summary:
Hello! This PR is a port of the [CSPRNG functions](http://php.net/csprng) `random_*()` which were [added in PHP7](https://wiki.php.net/rfc/easy_userland_csprng).

This PR is very much a work in progress and I'll need a lot of help from you smarter kids to get this implemented properly. But it works well on my machine thanks to some one-on-one help from jmikola.

The things that I know need to be addressed:

1. The PHP version [checks for sources of random at compile time](https://github.com/php/php-src/blob/97f159d70288ae94f9a05a370121bcbea760ed9a/Zend/Zend.m4#L406-L412). Not sure how to do this in HHVM.
2. PHP [supports Windows with `arc4random_buf()`](https://github.com/php/php-src/blob/cbcacbb2dad4c97ba5653f42729f7956193d9112/ext/standard/random.c#L86). Not sure if we want to support this as well as HHVM gets more Windows friendly... is that a thing?
3. There is [an open PR](https://github.com/php/php-src/pull/1397) to change the behavior of handling the cases when a reliable source of random cann

Closes https://github.com/facebook/hhvm/issues/5498
Closes https://github.com/facebook/hhvm/pull/5925

Reviewed By: sgolemon

Differential Revision: D2342229

Pulled By: jwatzman

fb-gh-sync-id: 419d949e3dd06fb1c13ecf86223c1b28eb57bb84