Update fbcode to use newer version of libdwarf
Summary:
Update fbcode to use the newer version of libdwarf (20150915) which has
been added to third-party2. This update is needed because the current version of
libdwarf does not handle TLS relocations.

There are only three users of libdwarf in fbcode: HHVM, symbolizer in folly, and
mobile/ios/atosl. Since the new version of libdwarf no longer provides the
dwarf_producer_init_b function, change the HHVM source to use the
dwarf_producer_init_c function instead (which is identical to the b
version). There was a define in mobile/ios/atosl which is now provided by
libdwarf directly, so that was removed.

(Making an attempt to add an appropriate reviewer for each affected project).

Reviewed By: Gownta

Differential Revision: D2713326

fb-gh-sync-id: a7ed19853e88ade71b6c92176dd2e7ecdebfbe5e